Create Kompwnd Investment with 1% ROI for 200days.
Join KOMPWND and hold KPW token for gain. In 200 days, you will get double of it and when the value of the token increase you will also gain.

You will gain double by earning 1% daily for 200 days.

Hold KPW for gain when the price increase. Because the crypto market is down now is the best time to buy and hold.

KOMPWND is a blockchain smart contract business that enables you to invest your KPW token in the smart contract and earn 1% daily to give you double in 200 days.


Minimum out of pocket compounding is $10 

Minimum out of profit compounding is $40. 

KPW is paired with ETH in the DEX exchange of Alacrity blockchain.
Minimum withdrawal amount is $50 if you are not compounding your earnings. So you can withdraw your profit and swap to ETH in the exchange and withdraw to your external wallet. 
Because KWP is paired with ETH when the price of ETH goes up it will also affect KPW.

JOIN NOW🚀🚀

First step:👇👇👇
Create Alacrity blockchain wallet here.
https://demo.alacritys.net

Second step:
Create KOMPWND account and connect to Alacrity  blockchain. 👇👇👇
http://kompwnd.io?referral=moneybag

Best Minimum to start is 20000 KPW

JOIN NOW
NO fear of volatility because you will earn double and still be able to compounding earnings.